feat: implement mobile-first Unified Agent Feed
- Updated `#triage-queue` instances to `#unified-agent-feed-section` to reflect the Unified Agent Feed Mobile MVP requirements.
- Replaced separate `test_agent_feed.spec.ts` and `test_viewport.spec.ts` with a cohesive, correctly formatted `dashboard-feed-mobile.spec.ts` mapping.
- Cleaned up `.agent-task` and output files.
